fix: resolve `map` and `groupBy` usage
